requirements.txt
streamlit
ccxt
pandas
numpy
google-generativeai
yfinance
